SK-1441 Fix Bug in Get method
Why
- A bug came in Get by unique values wherein, when multiple column values are passed, multiple column name keys are appended to request url.
- Due to new policy changes, any duplicate keys in request are not allowed which caused this scenario to fail and give errors.

Goal
- To fix this bug by changing logic in SDK for request URL creation.

Testing
- Additional unit tests are written to cover the new changes.
- Manual testing under different scenarios has been done.
